Twelve: Ensure we can navigate to target
Since callbacks are executed sequentially on the main thread
it can lead to issues when multiple user interactions happen
in quick succession like a double-click or a two-finger tap
on an item that triggers navigation to another fragment.

This happens because:
The first callback will navigate to a new fragment.
The second callback will attempt to navigate from
the already-replaced fragment to itself, crashing.

Change-Id: I2285e06594d3ba67cd32b6caf1fad5a0bd8647ca